pyproject.toml
[tool.poetry]
name = "drf-homeworks"
version = "0.1.0"
description = ""
authors = ["kifusss <kifus_06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
